use ooxml-schemas because http://poi.apache.org/help/faq.html#faq-N10025
There are two jar files available, as described in the components overview section. The full jar of all of the schemas is ooxml-schemas-1.4.jar (lower versions for older releases, see table below), and it is currently around 16mb. The smaller poi-ooxml-schemas jar is only about 6mb. This latter jar file only contains the typically used parts though.

Many users choose to use the smaller poi-ooxml-schemas jar to save space. However, the poi-ooxml-schemas jar only contains the XSDs and classes that are typically used, as identified by the unit tests. Every so often, you may try to use part of the file format which isn't included in the minimal poi-ooxml-schemas jar. In this case, you should switch to the full ooxml-schemas-1.4.jar. Longer term, you may also wish to submit a new unit test which uses the extra parts of the XSDs, so that a future poi-ooxml-schemas jar will include them.